+++ lib/backend/architecture/x86_64/x86_64Conv.ml
@@ -1753,6 +1753,41 @@ module Make (Spec : X86_64AsmSpec.X86_64AsmSpecification) = struct
               copy_result ~where ~register:rr9 ~rval_rktype rprogram
             in
             load_tag_instructions :: copy_instructions
+        | Array_len ->
+            let tte = List.hd parameters in
+            let array_len =
+              match tte.expr_rktype with
+              | RTArray { size; rktype = _ } ->
+                  size
+              | _ ->
+                  failwith "Weird: it should be an pointer array type"
+            in
+            let rr9 = Register.r9 in
+            let mov_instr =
+              Instruction
+                (Mov
+                   {
+                     size = iq;
+                     source = `ILitteral array_len;
+                     destination = `Register rr9;
+                   }
+                )
+            in
+            let copy_instructions =
+              copy_result ~where ~register:rr9 ~rval_rktype rprogram
+            in
+            mov_instr :: copy_instructions
+        | Array_ptr ->
+            let tte = List.hd parameters in
+            let rr9 = Register.r9 in
+            let _, instructions =
+              translate_tac_expression ~litterals ~target_dst:(`Register rr9)
+                rprogram fd tte
+            in
+            let copy_instructions =
+              copy_result ~where ~register:rr9 ~rval_rktype rprogram
+            in
+            instructions @ copy_instructions
         | Tos8
         | Tou8
         | Tos16